+++ src/transformer-components/SavedTransformerView.tsx
@@ -16,7 +16,7 @@ import {
   addInteractiveStateRequestListener,
   removeInteractiveStateRequestListener,
 } from "../utils/codapPhone/listeners";
-import "../ui-components/TransformerSaveUI.css";
+import "./TransformerSaveUI.css";
 import "./SavedTransformerView.css";
 import {
   useActiveTransformations,
